Rename some components to reflect better their roles
